bg(socialOauth): delivers changes on the social authentincation feature
