+++ .env.example
@@ -23,6 +23,7 @@ URL_DEV=https://warriorz.herokuapp.com
 
 API_BASE_URL = localhost:5000
 API_VERSION=v1
+FRONTEND_URL = https://barefoot-ui-staging.herokuapp.com
 
 RESET_URL=http://localhost:8000/login
 GOOGLE_ID = XXXXXXXXXXXXXXXXXXXXXXXXXXXX
