+++ package.json
@@ -11,7 +11,8 @@
       "server/services/email.create.js",
       "logger.js",
       "server/config/passport.js",
-      "server/helpers/profiler.js"
+      "server/helpers/profiler.js",
+      "server/controllers/socialOauth.js"
     ]
   },
   "scripts": {
@@ -60,7 +61,6 @@
     "morgan": "^1.10.0",
     "nodemailer": "^6.4.6",
     "nodemon": "^2.0.2",
-    "nyc": "^15.0.0",
     "passport": "^0.4.1",
     "passport-facebook": "^3.0.0",
     "passport-google-oauth20": "^2.0.0",
@@ -69,10 +69,8 @@
     "pg-hstore": "^2.3.3",
     "sequelize": "^5.21.5",
     "sequelize-cli": "^5.5.1",
-    "sinon": "^9.0.1",
     "socket.io": "^2.3.0",
     "socket.io-client": "^2.3.0",
-    "socketio-jwt": "^4.5.0",
     "swagger-jsdoc": "^4.0.0",
     "swagger-ui-express": "^4.1.4",
     "winston": "^3.2.1",
@@ -91,7 +89,6 @@
     "eslint-plugin-node": "^11.0.0",
     "eslint-plugin-prettier": "^3.1.2",
     "mocha-lcov-reporter": "^1.3.0",
-    "nyc": "^15.0.0",
-    "prettier": "^2.0.2"
+    "nyc": "^15.0.0"
   }
 }
